+++ client_code/popover.py
@@ -62,13 +62,9 @@ def popover(
     # We could be in the middle of a 'destroy'
     _wait_for_transition(popper_element)
 
-    if _has_popover(popper_element):
-        msg = (
-            "Warning: attempted to create a popover on a component that already has one. This will have no effect.\n"
-            "Destroy the popover before creating a new one using component.pop('destroy').\n"
-            "Or, use has_popover() to check if this component aleady has a popover."
-        )
-        print(msg)
+    has_popover = _has_popover(popper_element)
+    _check_warnings(has_popover, component)
+    if has_popover:
         # return here since adding a new popover has no effect
         return
 
@@ -192,6 +188,32 @@ def _set_data(popper_element, attr, value):
         data[attr] = value
 
 
+_warnings = {}
+
+
+def _check_warnings(has_popover, component):
+    if has_popover and not _warnings.get("has_pop"):
+        msg = (
+            "Warning: attempted to create a popover on a component that already has one. This will have no effect.\n"
+            "Destroy the popover before creating a new one using component.pop('destroy').\n"
+            "Or, use has_popover() to check if this component aleady has a popover before creating a new one."
+        )
+        print(msg)
+        _warnings["has_pop"] = True
+    elif (
+        component is not None
+        and component.parent is not None
+        and type(component.parent) is not _anvil.Container
+        and not _warnings.get("has_parent")
+    ):
+        # if the parent is an anvil Container then it's probably an internal issue
+        print(
+            "Warning: the popover content already has a parent this can cause unusual behaviour.\n"
+            "Support for this may be removed in a future version."
+        )
+        _warnings["has_parent"] = True
+
+
 def _add_transition_behaviour(component, popper_element, popper_id):
     # clean up our previous event handlers
     popper_element.off(
@@ -206,20 +228,22 @@ def _add_transition_behaviour(component, popper_element, popper_id):
             _set_data(popper_element, "inTransition", None)
             popper_element.off("shown.bs.popover", f)
             resolve(None)
+            if component is None:
+                return
+            open_form = _anvil.get_open_form()
+            if open_form is not None and fake_container.parent is None:
+                open_form.add_component(fake_container)
+            if component.parent is None:
+                # we add the component to a Container component
+                # this doesn't really add it to the dom
+                # it just allows us to use anvil's underlying show hide architecture
+                fake_container.add_component(component)
 
         popper_element.on("shown.bs.popover", f)
 
     def show_in_transition(e):
         _set_data(popper_element, "inTransition", _Promise(resolve_shown))
         _visible_popovers[popper_id] = popper_element
-        open_form = _anvil.get_open_form()
-        if open_form is not None and fake_container.parent is None:
-            open_form.add_component(fake_container)
-        if component is not None and component.parent is None:
-            # we add the component to a Container component
-            # this doesn't really add it to the dom
-            # it just allows us to use anvil's underlying show hide architecture
-            fake_container.add_component(component)
 
     popper_element.on("show.bs.popover", show_in_transition)
 
@@ -298,9 +322,13 @@ _window.addEventListener("resize", _update_positions)
 _scrolling = False
 
 
-def _hide_on_scroll(*e):
+def _hide_on_scroll(e):
     global _scrolling
-    if _scrolling or not _visible_popovers:
+    if (
+        _scrolling
+        or not _visible_popovers
+        or e.target.closest(".anvil-popover") is not None
+    ):
         return
 
     _scrolling = True