+++ CHANGELOG.md
@@ -26,6 +26,8 @@
   https://github.com/anvilistas/anvil-extras/pull/199
 * Using routing load_from_cache=False to reload the current form works correctly
   https://github.com/anvilistas/anvil-extras/issues/243
+* PageBreak retains its thickness in print mode
+  https://github.com/anvilistas/anvil-extras/issues/263
 
 # v1.8.1 14-Oct-2021