+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * `utils.correct_canvas_resolution()` - canvas elements can look blurry on retina displays
   This function sharpens the resolution of a canvas element when called in the reset event
   https://github.com/anvilistas/anvil-extras/pull/202
+* Slider - visual properties - `handle_size`, `bar_height`, and `role` added
+  https://github.com/anvilistas/anvil-extras/pull/261
 
 ## Updates:
 * storage supports `datetime` and `date` objects